// prototype-v1/src/combat/types.ts
export interface InfoNugget {
  id: string;
  name: string;
  longDescription: string;
  imageUrl?: string;
  defaultCardId?: string;
}